+++ src/tradelayer/tradelayer.h
@@ -543,7 +543,7 @@ namespace mastercore
 
   int64_t sum_check_upnl(const std::string& address); //  sum of all upnls for a given address.
 
-  int64_t pos_margin(uint32_t contractId, const std::string& address, uint32_t margin_requirement); // return mainteinance margin for a given contrand and address
+  int64_t pos_margin(uint32_t contractId, const std::string& address, uint64_t margin_requirement); // return mainteinance margin for a given contrand and address
 
   bool makeWithdrawals(int Block); // make the withdrawals for multisig channels
